requirements.txt
Flask==2.3.2
kafka-python==2.0.2
keras==2.12.0
matplotlib-inline==0.1.3
numpy==1.23.5
pandas==1.5.2
pyspark==3.4.0
requests==2.31.0
selenium==4.7.2
sklearn==0.0.post4
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0